backend/pyproject.toml
[project]
name = "aibia-backend"
version = "0.1.0"
description = "AI agent for small businesses"
authors = [
    {name = "joesaiagent",email = "joesaiagent@gmail.com"}
]
requires-python = "^3.14"
dependencies = [
    "anthropic (>=0.86.0,<0.87.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "tavily-python (>=0.7.23,<0.8.0)",
    "google-auth (>=2.49.1,<3.0.0)",
    "google-auth-oauthlib (>=1.3.0,<2.0.0)",
    "google-api-python-client (>=2.193.0,<3.0.0)",
    "msal (>=1.35.1,<2.0.0)",
    "cryptography (>=46.0.5,<47.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
